refactor: 딜러와 플레이어 로직 추상화
