+++ src/main/java/domain/Dealer.java
@@ -1,23 +1,9 @@
 package domain;
 
-public class Dealer {
-
-    private final Hand hand;
-
-    public Dealer() {
-        this.hand = new Hand();
-    }
-
-    public void hit(Card card) {
-        hand.drawCard(card);
-    }
+public class Dealer extends Participant {
 
     public Card getOpenCard() {
         return hand.getHand().getFirst();
-    }
-
-    public Hand getHand() {
-        return hand;
     }
 
     public boolean shouldHit() {
